+++ cpu/heat2d.py
@@ -99,11 +99,11 @@ def resolve(
     Stability requires lambda_x + lambda_y <= 0.5.
     """
     # Convert inputs
-    u0       = np.asarray(u0, float)
-    g_left   = np.asarray(g_left, float)
-    g_right  = np.asarray(g_right, float)
-    g_bottom = np.asarray(g_bottom, float)
-    g_top    = np.asarray(g_top, float)
+    u0       = np.asarray(u0, np.float32)
+    g_left   = np.asarray(g_left, np.float32)
+    g_right  = np.asarray(g_right, np.float32)
+    g_bottom = np.asarray(g_bottom, np.float32)
+    g_top    = np.asarray(g_top, np.float32)
     a, b = interval_x
     c, d = interval_y
 
@@ -127,7 +127,7 @@ def resolve(
         )
 
     # Initial conditions
-    u = np.zeros((n_x, n_y, n_t), dtype=float)
+    u = np.zeros((n_x, n_y, n_t), dtype=np.float32)
     u[:, :, 0] = u0
 
     # Boundary conditions
